+++ addons/auth_totp/models/__init__.py
@@ -2,5 +2,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import auth_totp
+from . import auth_totp_rate_limit_log
 from . import res_users
 from . import totp
